% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateStripes.R
\name{climateStripes}
\alias{climateStripes}
\alias{ClimateStripes}
\alias{climatestripes}
\title{climateStripes}
\usage{
climateStripes(
  city.name = NULL,
  weather.station.id = NULL,
  startyear.mean = 1961,
  endyear.mean = 1990,
  plot.what = "all",
  style = "continuous"
)
}
\arguments{
\item{city.name}{A character}

\item{weather.station.id}{A character}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{plot.what}{A character (can be warmingstripes, precipitation,
sunlight, windspeed)}

\item{style}{A character (could be for
plot.what == temperature: continuous or ...)}
}
\description{
Creates a climate plots that show annual deviation from a mean value
}
\details{
This main function of the package fetches the data and calls the
required plotting functions.
}
\author{
Kai Budde
}
