% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherData.R
\name{getWeatherData}
\alias{getWeatherData}
\alias{getweatherdata}
\alias{getWeatherdata}
\alias{getweatherData}
\title{getWeatherData}
\usage{
getWeatherData(
  link.to.data = NULL,
  weather.station.id = NULL,
  filename.data = NULL
)
}
\arguments{
\item{link.to.data}{A character (ftp link to weather data)}

\item{weather.station.id}{A character (5 digit station number)}

\item{filename.data}{A character (file name that contains the data)}
}
\description{
Downloads weather data of specific DWD waeather station
}
\details{
This function downloads all weather data of a specific weather
station from the servers of the Deutscher Wetterdienst (DWD)
}
\author{
Kai Budde
}
