% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherStations.R
\name{getWeatherStations}
\alias{getWeatherStations}
\alias{getweatherstations}
\title{getWeatherStations}
\usage{
getWeatherStations()
}
\description{
Downloads information about all weather stations of DWD
}
\details{
This function downloads information about all weather stations
of Deutscher Wetterdienst (DWD) and saves the information in a data frame
}
\author{
Kai Budde
}
