% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrecipitation.R
\name{plotPrecipitation}
\alias{plotPrecipitation}
\alias{plotwarmingstripes}
\alias{plotWarmingstripes}
\alias{plotwarmingStripes}
\title{plotPrecipitation}
\usage{
plotPrecipitation(
  df.temp,
  startyear.mean,
  endyear.mean,
  style = "continuous",
  station.name = NULL
)
}
\arguments{
\item{df.temp}{A data frame}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{style}{A character}

\item{station.name}{a character}
}
\description{
Plot the mean temperature data
}
\details{
This function takes a data frame with daily temperature data for
given station, calculates annual mean and plots it.
}
\author{
Kai Budde
}
