% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrecipitationStripes.R
\name{plotPrecipitationStripes}
\alias{plotPrecipitationStripes}
\alias{plotprecipitationstripes}
\alias{plotprecipitationStripes}
\alias{plotPrecipitationstripes}
\title{plotPrecipitationStripes}
\usage{
plotPrecipitationStripes(
  df,
  startyear.mean,
  endyear.mean,
  style = "continuous",
  station.name = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{style}{A character}

\item{station.name}{a character}
}
\description{
Plot the deviation of cumulative precipitation of a year
}
\details{
This function takes a data frame with daily precipitation (in mm)
for a given station, calculates annual sum and the deviation from a
reference period and plots the deviations as color coded bars.
}
\author{
Kai Budde
}
