% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSunhoursStripes.R
\name{plotSunhoursStripes}
\alias{plotSunhoursStripes}
\alias{plotsunhoursstripes}
\alias{plotSunhoursstripes}
\alias{plotsunhoursgStripes}
\title{plotSunhoursStripes}
\usage{
plotSunhoursStripes(
  df,
  startyear.mean,
  endyear.mean,
  style = "continuous",
  station.name = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{style}{A character}

\item{station.name}{a character}
}
\description{
Plot the deviation of the mean daily sunshine duration
}
\details{
This function takes a data frame with daily sunshine duration
data for a given station, calculates annual mean and the deviation from a
reference period and plots the deviations a color coded bars.
}
\author{
Kai Budde
}
