% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.stripes.r
\name{rings.breakpoint.f}
\alias{rings.breakpoint.f}
\title{Superimpose a breakpoint circle on a climate rings plot}
\usage{
rings.breakpoint.f(
  time.vector,
  temperature.vector,
  border.colour = "yellow",
  lty = 2,
  lwd = 5,
  nv = 1000
)
}
\arguments{
\item{time.vector}{the time series vector, must be the same as used in
\code{climate.col.rings.f}}

\item{temperature.vector}{a time series of any environmental variable, must be the same as
used in \code{climate.col.rings.f}}

\item{border.colour}{colour of the breakpoint circle border. Default "yellow"}

\item{lty}{line type for the breakpoint circle border. Default 2 (dashed)}

\item{lwd}{line width for the breakpoint circle border. Default 5}

\item{nv}{number of vertices used to draw the circle. Default 1000}
}
\description{
Fits a segmented (piecewise) linear regression to the temperature time series
  and superimposes a dashed circle on the climate rings plot at the radius corresponding to
  the estimated breakpoint year. The breakpoint year is returned invisibly and printed to
  the console.
}
\details{
The radius scale matches that used internally by \code{climate.col.rings.f}: time
  index rescaled to 0.1--5 so that the newest year has the largest radius.
}
\seealso{
climate.col.rings.f segmented::segmented plotrix::draw.circle
}
