% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.stripes.r
\name{superimpose.data.f}
\alias{superimpose.data.f}
\title{Superimpose the data series and a smooth gam (spline) trend line on the climate stripe plot}
\usage{
superimpose.data.f(
  time.vector,
  temperature.vector,
  data.colour = "yellow",
  spline = T,
  spline.colour = "white",
  ...
)
}
\arguments{
\item{time.vector}{the time series vector}

\item{temperature.vector}{a times series of any environmental variable (climate stripes have been developed for temperature) whose values correspond to the time vector}

\item{data.colour}{colour of the data line to superimpose}

\item{spline}{fit a spline (gam) and superimpose in addition to the data}

\item{spline.colour}{colour of the spline line to superimpose}

\item{...}{additional arguments to the trend line that "lines" will accept, see par}
}
\description{
Superimpose data and a gam smooth (spline) on the climate stripes plot. It can be visually more striking and can be more science friendly while still conveying the main colour temperature message of the climate stripes plot
}
\details{
Superimposes data after rescaling from 0 to 1, i.e. the lowest point in the series will be at the bottom of the graph and the highest at the top. The
      gam is calculated using mgcv default for gam
}
\seealso{
lines par mgcv::gam smooth.spline
}
