% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateStripes.R
\name{ClimateStripes}
\alias{ClimateStripes}
\alias{warmingstripes}
\title{climateStripes}
\usage{
ClimateStripes(city.name = NULL, weather.station.id = NULL,
  startyear.mean = 1961, endyear.mean = 1990, plot.what = "all",
  style = "continuous")
}
\arguments{
\item{city.name}{A character}

\item{weather.station.id}{A character}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{plot.what}{A character (can be warmingstripes, precipitation,
sunlight, windspeed)}

\item{style}{A character (could be for
plot.what == temperature: continuous or ...)}
}
\description{
Create a plot with annual deviation from mean temperature
}
\details{
This main function of the package creates a barplot where every
colored bar represents the deviation of the annual mean temperature from
a 30-years mean. Dark red colors stand for higher mean temperatures and
dark blue colors for lower mean temperatures.
}
\author{
Kai Budde
}
