% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.stripes.r
\name{colour.gradient.legend.f}
\alias{colour.gradient.legend.f}
\title{Draws a gradient legend for a climate stripes plot}
\usage{
colour.gradient.legend.f(
  xleft,
  ybottom,
  xright,
  ytop,
  colour.vec = c("blue", "red"),
  ncolours = 500,
  labels = T,
  var.min.label,
  var.max.label,
  text.col.legend
)
}
\arguments{
\item{xleft}{the lower left x coordinate of the legend}

\item{ybottom}{the lower left y coordinate of the legend}

\item{xright}{the upper right x coordinate of the legend}

\item{ytop}{the upper right y coordinate of the legend}

\item{colour.vec}{a vector of colours to ramp between (see colorRampPalette, "colors")}

\item{ncolours}{the number of colours to show on the legend. Default is 500, a large number produces a smooth legend colour gradient}

\item{labels}{default TRUE. If TRUE then values you supply (var.min.label) and (var.max.label) are shown on the legend}

\item{var.min.label}{the value for the highest temperature to show on the legend}

\item{var.max.label}{the value for the lowest temperature to show on the legend}

\item{text.col.legend}{colour of the legend text}
}
\description{
This draws a colour gradient legend for use as a climate stripes legend. It uses rect to draw many rectangles (500) and colours
      each with a ramp palette so that it appears as a smooth gradient. It is also possible to do this as a rasterImage but these images
      can be hard to work with and do not scale properly especially for multiplot layouts.
}
\seealso{
rect colorRampPalette
}
