% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWarmingStripes.R
\name{plotWarmingStripes}
\alias{plotWarmingStripes}
\alias{plotwarmingstripes}
\title{plotWarmingStripes}
\usage{
plotWarmingStripes(df.temp, startyear.mean, endyear.mean,
  style = "continuous")
}
\arguments{
\item{df.temp}{A data frame}

\item{endyear.mean}{A number}

\item{staryear.mean}{A number}
}
\description{
Plot the mean temperature data
}
\details{
This function takes a data frame with daily temperature data for
given station, calculates annual mean and plots it.
}
\author{
Kai Budde
}
