% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWarmingStripes.R
\name{plotWarmingStripes}
\alias{plotWarmingStripes}
\alias{plotwarmingstripes}
\alias{plotWarmingstripes}
\alias{plotwarmingStripes}
\title{plotWarmingStripes}
\usage{
plotWarmingStripes(
  df,
  startyear.mean,
  endyear.mean,
  style = "continuous",
  station.name = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}

\item{style}{A character}

\item{station.name}{a character}
}
\description{
Plot the deviation of mean temperature per year
}
\details{
This function takes a data frame with daily temperature data for
given station, calculates annual mean and the deviation from a
reference period and plots the deviations a color coded bars.


This main function of the package creates a barplot where every
colored bar represents the deviation of the annual mean temperature from
a 30-years mean. Dark red colors stand for higher mean temperatures and
dark blue colors for lower mean temperatures.
}
\author{
Kai Budde
}
