% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmingStripes.R
\name{warmingStripes}
\alias{warmingStripes}
\alias{warmingstripes}
\title{warmingStripes}
\usage{
warmingStripes(input_file, startyear.mean = 1961, endyear.mean = 1990)
}
\arguments{
\item{input_file}{A character}

\item{startyear.mean}{A number}

\item{endyear.mean}{A number}
}
\description{
Create a plot with annual deviation from mean temperature
}
\details{
The package creats a barplot where every colored bar represents
the deviation of the annual mean temperature from a 30-years mean. Dark
red colors stand for higher mean temperatures and dark blue colors for
lower mean temperatures.
}
\author{
Kai Budde
}
